+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Deprecated `Gos\Bundle\WebSocketBundle\Pusher\PusherInterface` and `Gos\Bundle\WebSocketBundle\Pusher\ServerPushHandlerInterface`, and all related services, in favor of the Symfony Messenger component
 - Removed `Gos\Bundle\WebSocketBundle\Client\ClientStorageInterface::setStorageDriver()`, this method should no longer be relied on
 - [MINOR B/C BREAK] Changed the (final) `Gos\Bundle\WebSocketBundle\Client\ClientStorage` constructor to require a `Gos\Bundle\WebSocketBundle\Client\Driver\DriverInterface` instance as the first argument, this only affects users manually instantiating an instance of the storage class
+- Deprecated unused `gos_web_socket.client.storage.prefix` configuration node and container parameter
 
 ## 3.0.0 (2020-04-02)