+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Removed `Gos\Bundle\WebSocketBundle\Client\ClientStorageInterface::setStorageDriver()`, this method should no longer be relied on
 - [MINOR B/C BREAK] Changed the (final) `Gos\Bundle\WebSocketBundle\Client\ClientStorage` constructor to require a `Gos\Bundle\WebSocketBundle\Client\Driver\DriverInterface` instance as the first argument, this only affects users manually instantiating an instance of the storage class
 - Deprecated unused `gos_web_socket.client.storage.prefix` configuration node and container parameter
+- Address deprecations in marking configuration nodes, services, and service aliases deprecated in Symfony 5.1
 
 ## 3.0.0 (2020-04-02)