+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 3.1.0 (2020-??-??)
+## 3.1.0 (2020-05-31)
 
 - Use the `symfony/deprecation-contracts` package to trigger runtime deprecation notices
 - Deprecated `Gos\Bundle\WebSocketBundle\Pusher\PusherInterface` and `Gos\Bundle\WebSocketBundle\Pusher\ServerPushHandlerInterface`, and all related services, in favor of the Symfony Messenger component